# backend/gen/blog_api.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// CreatePostRequest defines model for CreatePostRequest.
type CreatePostRequest struct {
	Body  string `json:"body"`
	Title string `json:"title"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message string `json:"message"`
}

// Post defines model for Post.
type Post struct {
	Body      string    `json:"body"`
	CreatedAt time.Time `json:"createdAt"`
	Id        string    `json:"id"`
	Title     string    `json:"title"`
	UpdatedAt time.Time `json:"updatedAt"`
}

// PostList defines model for PostList.
type PostList struct {
	Posts []Post `json:"posts"`
	Total int32  `json:"total"`
}

// UpdatePostRequest defines model for UpdatePostRequest.
type UpdatePostRequest struct {
	Body  *string `json:"body,omitempty"`
	Title *string `json:"title,omitempty"`
}

// PostsList200JSONResponseBody defines parameters for PostsList.
type PostsList200JSONResponseBody struct {
	union json.RawMessage
}

// PostsCreateJSONRequestBody defines body for PostsCreate for application/json ContentType.
type PostsCreateJSONRequestBody = CreatePostRequest

// PostsUpdateJSONRequestBody defines body for PostsUpdate for application/json ContentType.
type PostsUpdateJSONRequestBody = UpdatePostRequest

// AsPostList returns the union data inside the PostsList200JSONResponseBody as a PostList
func (t PostsList200JSONResponseBody) AsPostList() (PostList, error) {
	var body PostList
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromPostList overwrites any union data inside the PostsList200JSONResponseBody as the provided PostList
func (t *PostsList200JSONResponseBody) FromPostList(v PostList) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergePostList performs a merge with any union data inside the PostsList200JSONResponseBody, using the provided PostList
func (t *PostsList200JSONResponseBody) MergePostList(v PostList) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsErrorResponse returns the union data inside the PostsList200JSONResponseBody as a ErrorResponse
func (t PostsList200JSONResponseBody) AsErrorResponse() (ErrorResponse, error) {
	var body ErrorResponse
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromErrorResponse overwrites any union data inside the PostsList200JSONResponseBody as the provided ErrorResponse
func (t *PostsList200JSONResponseBody) FromErrorResponse(v ErrorResponse) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeErrorResponse performs a merge with any union data inside the PostsList200JSONResponseBody, using the provided ErrorResponse
func (t *PostsList200JSONResponseBody) MergeErrorResponse(v ErrorResponse) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t PostsList200JSONResponseBody) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *PostsList200JSONResponseBody) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /health)
	HealthCheck(c *gin.Context)

	// (GET /posts)
	PostsList(c *gin.Context)

	// (POST /posts)
	PostsCreate(c *gin.Context)

	// (DELETE /posts/{id})
	PostsDelete(c *gin.Context, id string)

	// (GET /posts/{id})
	PostsGet(c *gin.Context, id string)

	// (PUT /posts/{id})
	PostsUpdate(c *gin.Context, id string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.HealthCheck(c)
}

// PostsList operation middleware
func (siw *ServerInterfaceWrapper) PostsList(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostsList(c)
}

// PostsCreate operation middleware
func (siw *ServerInterfaceWrapper) PostsCreate(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostsCreate(c)
}

// PostsDelete operation middleware
func (siw *ServerInterfaceWrapper) PostsDelete(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostsDelete(c, id)
}

// PostsGet operation middleware
func (siw *ServerInterfaceWrapper) PostsGet(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostsGet(c, id)
}

// PostsUpdate operation middleware
func (siw *ServerInterfaceWrapper) PostsUpdate(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostsUpdate(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/health", wrapper.HealthCheck)
	router.GET(options.BaseURL+"/posts", wrapper.PostsList)
	router.POST(options.BaseURL+"/posts", wrapper.PostsCreate)
	router.DELETE(options.BaseURL+"/posts/:id", wrapper.PostsDelete)
	router.GET(options.BaseURL+"/posts/:id", wrapper.PostsGet)
	router.PUT(options.BaseURL+"/posts/:id", wrapper.PostsUpdate)
}

type HealthCheckRequestObject struct {
}

type HealthCheckResponseObject interface {
	VisitHealthCheckResponse(w http.ResponseWriter) error
}

type HealthCheck200JSONResponse struct {
	Status string `json:"status"`
}

func (response HealthCheck200JSONResponse) VisitHealthCheckResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostsListRequestObject struct {
}

type PostsListResponseObject interface {
	VisitPostsListResponse(w http.ResponseWriter) error
}

type PostsList200JSONResponse = PostsList200JSONResponseBody

func (response PostsList200JSONResponse) VisitPostsListResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.union); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostsCreateRequestObject struct {
	Body *PostsCreateJSONRequestBody
}

type PostsCreateResponseObject interface {
	VisitPostsCreateResponse(w http.ResponseWriter) error
}

type PostsCreate200JSONResponse ErrorResponse

func (response PostsCreate200JSONResponse) VisitPostsCreateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostsCreate201JSONResponse Post

func (response PostsCreate201JSONResponse) VisitPostsCreateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type PostsDeleteRequestObject struct {
	Id string `json:"id"`
}

type PostsDeleteResponseObject interface {
	VisitPostsDeleteResponse(w http.ResponseWriter) error
}

type PostsDelete204Response struct {
}

func (response PostsDelete204Response) VisitPostsDeleteResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type PostsDelete404JSONResponse ErrorResponse

func (response PostsDelete404JSONResponse) VisitPostsDeleteResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type PostsGetRequestObject struct {
	Id string `json:"id"`
}

type PostsGetResponseObject interface {
	VisitPostsGetResponse(w http.ResponseWriter) error
}

type PostsGet200JSONResponse Post

func (response PostsGet200JSONResponse) VisitPostsGetResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostsGet404JSONResponse ErrorResponse

func (response PostsGet404JSONResponse) VisitPostsGetResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type PostsUpdateRequestObject struct {
	Id   string `json:"id"`
	Body *PostsUpdateJSONRequestBody
}

type PostsUpdateResponseObject interface {
	VisitPostsUpdateResponse(w http.ResponseWriter) error
}

type PostsUpdate200JSONResponse Post

func (response PostsUpdate200JSONResponse) VisitPostsUpdateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostsUpdate404JSONResponse ErrorResponse

func (response PostsUpdate404JSONResponse) VisitPostsUpdateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /health)
	HealthCheck(ctx context.Context, request HealthCheckRequestObject) (HealthCheckResponseObject, error)

	// (GET /posts)
	PostsList(ctx context.Context, request PostsListRequestObject) (PostsListResponseObject, error)

	// (POST /posts)
	PostsCreate(ctx context.Context, request PostsCreateRequestObject) (PostsCreateResponseObject, error)

	// (DELETE /posts/{id})
	PostsDelete(ctx context.Context, request PostsDeleteRequestObject) (PostsDeleteResponseObject, error)

	// (GET /posts/{id})
	PostsGet(ctx context.Context, request PostsGetRequestObject) (PostsGetResponseObject, error)

	// (PUT /posts/{id})
	PostsUpdate(ctx context.Context, request PostsUpdateRequestObject) (PostsUpdateResponseObject, error)
}

type StrictHandlerFunc func(ctx *gin.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictGinServerOptions struct {
	// RequestErrorHandlerFunc is called when a request cannot be parsed or
	// decoded. It is invoked for JSON bind failures, form parse/bind errors,
	// multipart reader errors, media type parse errors, missing multipart
	// boundaries, and request body read errors. The default returns 400.
	RequestErrorHandlerFunc func(ctx *gin.Context, err error)
	// HandlerErrorFunc is called when the application handler (or any
	// middleware wrapping it) returns a non-nil error. The default returns 500.
	HandlerErrorFunc func(ctx *gin.Context, err error)
	// ResponseErrorHandlerFunc is called when the response object fails to
	// serialize (Visit*Response returns an error) or when the handler returns
	// an unexpected response type. The default returns 500.
	ResponseErrorHandlerFunc func(ctx *gin.Context, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictGinServerOptions{
		RequestErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		},
		HandlerErrorFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
		ResponseErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictGinServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		}
	}
	if options.HandlerErrorFunc == nil {
		options.HandlerErrorFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictGinServerOptions
}

// HealthCheck operation middleware
func (sh *strictHandler) HealthCheck(ctx *gin.Context) {
	var request HealthCheckRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.HealthCheck(ctx, request.(HealthCheckRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "HealthCheck")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(HealthCheckResponseObject); ok {
		if err := validResponse.VisitHealthCheckResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostsList operation middleware
func (sh *strictHandler) PostsList(ctx *gin.Context) {
	var request PostsListRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostsList(ctx, request.(PostsListRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostsList")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(PostsListResponseObject); ok {
		if err := validResponse.VisitPostsListResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostsCreate operation middleware
func (sh *strictHandler) PostsCreate(ctx *gin.Context) {
	var request PostsCreateRequestObject

	var body PostsCreateJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(ctx, err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostsCreate(ctx, request.(PostsCreateRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostsCreate")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(PostsCreateResponseObject); ok {
		if err := validResponse.VisitPostsCreateResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostsDelete operation middleware
func (sh *strictHandler) PostsDelete(ctx *gin.Context, id string) {
	var request PostsDeleteRequestObject

	request.Id = id

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostsDelete(ctx, request.(PostsDeleteRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostsDelete")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(PostsDeleteResponseObject); ok {
		if err := validResponse.VisitPostsDeleteResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostsGet operation middleware
func (sh *strictHandler) PostsGet(ctx *gin.Context, id string) {
	var request PostsGetRequestObject

	request.Id = id

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostsGet(ctx, request.(PostsGetRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostsGet")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(PostsGetResponseObject); ok {
		if err := validResponse.VisitPostsGetResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostsUpdate operation middleware
func (sh *strictHandler) PostsUpdate(ctx *gin.Context, id string) {
	var request PostsUpdateRequestObject

	request.Id = id

	var body PostsUpdateJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(ctx, err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostsUpdate(ctx, request.(PostsUpdateRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostsUpdate")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(PostsUpdateResponseObject); ok {
		if err := validResponse.VisitPostsUpdateResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}